spark-processor/requirements.txt
pyspark==3.3.1
kafka-python==2.0.2
requests==2.31.0
psycopg2-binary==2.9.9
pandas==2.0.3
sqlalchemy==2.0.23
